Replaced xrange with range
In Python 3, xrange was renamed to range (more or less). Replacing
current calls to xrange with range has almost no affect on memory or
performance as the size of the resulting lists being returned are tiny
in the context. That is, replacing xrange(1000000000000) with range
would use a lot of RAM, but replacing xrange(10) with range will be
almost unmeasurable.